Added AuthorizationPrincipal, changed SecurityConfig

+++ collegecommerce-server/src/main/java/learn/capstone/controllers/ListingController.java
@@ -4,15 +4,16 @@ import learn.capstone.domain.ListingResult;
 import learn.capstone.domain.ListingService;
 import learn.capstone.domain.ResultType;
 import learn.capstone.models.Listing;
+import learn.capstone.models.User;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
 @RestController
 @RequestMapping("/api/listing")
-@CrossOrigin(origins = {"http://localhost:3000"})
 public class ListingController {
     private final ListingService service;
     private ResponseEntity<Object> objectResponseEntity;
@@ -32,7 +33,10 @@ public class ListingController {
     }
 
     @PostMapping
-    public ResponseEntity<?> create(@RequestBody Listing listing) {
+    public ResponseEntity<?> create(@AuthenticationPrincipal User authUser, @RequestBody Listing listing) {
+        if (authUser.getUserId() != listing.getUserId()) {
+            return new ResponseEntity<>(HttpStatus.CONFLICT);
+        }
         ListingResult result = service.create(listing);
         if (!result.isSuccess()) {
             return new ResponseEntity<>(result.getErrorMessages(), HttpStatus.BAD_REQUEST); //400
@@ -41,9 +45,10 @@ public class ListingController {
     }
 
     @PutMapping("/{listingId}")
-    public ResponseEntity<Object> update(@PathVariable int listingId, @RequestBody Listing listing) {
-        System.out.println(listingId);
-        System.out.println(listing.getListingId());
+    public ResponseEntity<Object> update(@AuthenticationPrincipal User authUser, @PathVariable int listingId, @RequestBody Listing listing) {
+        if (authUser.getUserId() != listing.getUserId()) {
+            return new ResponseEntity<>(HttpStatus.CONFLICT);
+        }
         if (listingId != listing.getListingId()) {
             return new ResponseEntity<>(HttpStatus.CONFLICT); // 409
         }
@@ -60,7 +65,17 @@ public class ListingController {
     }
 
     @DeleteMapping("/{listingId}")
-    public ResponseEntity<Void> delete(@PathVariable int listingId) {
+    public ResponseEntity<Void> delete(@AuthenticationPrincipal User authUser, @PathVariable int listingId) {
+        Listing listing = service.findByListingId(listingId);
+        System.out.println(listing);
+        if (listing == null) {
+            System.out.println("Listing was null");
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        if (listing.getUserId() != authUser.getUserId()) {
+            return new ResponseEntity<>(HttpStatus.CONFLICT);
+        }
+
         ListingResult result = service.deleteByListingId(listingId);
         if (result.getResultType() == ResultType.NOT_FOUND) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND); //404
